Added a class that boots arcadeflex with no video output and no info popup.
MainStream class can be used to start the emulation and query the class from screenshots.
The implementation is really bulky and experimental.
Next steps would be to actually make arcadeflex capable to run safely as a thread that can be stopped gracefully.
Also sound has to be somehow muted and extracted.
(Changes are fewer than shown due to code formatting)